feat(cli): the retrieval knobs reach the settings block that exists to hold them
Deliverable 8 of #712, plus the doc bookkeeping the first seven owed.

Eight ranking constants and two per-query budgets were hard-coded, while a
`context.*` settings block sat in the tree specifically to hold them —
"entirely inert in Phase 0", as its own module doc said. `context.retrieval`
is now live, and live regardless of `lifecycle.enabled`: it configures the
retrieval plane that already runs for every user, not the adaptive loop.

Every default is the value that shipped, asserted against the `DEFAULT_*`
constants themselves rather than re-typed — so "keeping current values as
defaults" is checked, not claimed. Out-of-range values are clamped by
`RecallTuning::sanitized`, never rejected: this is a file a person edits, so
zero and negative are reachable, and failing a turn over a typo in a tuning
knob is a worse answer than ignoring it. A zero shortlist multiple would
make every recall empty; a zero `coverage_topk` divides by zero; a
non-positive fusion constant inverts the ranking.

The settings read degrades to defaults, deliberately unlike the suppression
read next to it, which fails closed. The degraded outcomes are not
comparable: a slightly different frame order versus surfacing something a
person asked to forget.

Witness: widening `mmr_candidate_multiple` widens `considered`, and a
coverage floor of 1.0 forces the labeled lexical fallback — a settings block
that deserializes but steers nothing is what this replaces.

Docs the earlier commits invalidated, corrected here rather than left:
ADR 0003's recall-layer gap is recorded closed (the issue names this), the
spec's defect table gets a status line, and `stella-context/README.md`'s
"Nothing forgets" / "every live node becomes a candidate" / "forget does not
reach in here" caveats are rewritten — all three were true when written and
are now false. Its file map is updated for the split.

`memory.rs`'s tests move to a sibling module (pure move), dropping it from
1,654 to 929 lines and retiring its baseline entry; the settings read lands
in `memory/tuning.rs`.

`make gate` green. Closes the deliverable list of #712 except the stale
strings.
