+++ scripts/file-size-baseline.txt
@@ -12,7 +12,6 @@
 1599 stella-cli/src/candidate_ws.rs
 4465 stella-cli/src/command_deck.rs
 1790 stella-cli/src/main.rs
-1643 stella-cli/src/memory.rs
 2095 stella-core/src/bus.rs
 2252 stella-core/src/driver.rs
 2818 stella-core/src/driver/tests.rs
