+++ stella-cli/src/settings/context.rs
@@ -34,6 +34,7 @@ pub struct ContextSettings {
     pub promotion: PromotionSettings,
     pub efficacy: EfficacySettings,
     pub retention: RetentionSettings,
+    pub retrieval: RetrievalSettings,
 }
 
 /// Master switch for the whole adaptive-context lifecycle.
@@ -183,6 +184,94 @@ impl Default for RetentionSettings {
     }
 }
 
+/// How recall ranks, diversifies, and budgets — the knobs that were `const`s.
+///
+/// Unlike the rest of this block, these are **live**, and they are live
+/// regardless of [`LifecycleSettings::enabled`]: they configure the retrieval
+/// plane that already runs for every user, not the adaptive loop. Every default
+/// is exactly the value that shipped hard-coded, so a workspace that configures
+/// nothing behaves identically (#712 deliverable 8).
+///
+/// Out-of-range values are clamped by `stella_context::RecallTuning::sanitized`
+/// rather than rejected. This is a file a person edits, so the invalid values
+/// are reachable, and failing a turn over a typo in a tuning knob is a worse
+/// answer than ignoring it.
+#[derive(Debug, Clone, PartialEq, Serialize, Deserialize)]
+#[serde(default)]
+pub struct RetrievalSettings {
+    /// How many frames a prompt recall asks for.
+    pub max_frames: u32,
+    /// The token budget those frames are packed into.
+    pub max_tokens: u32,
+    /// Reciprocal-rank-fusion constant. Higher flattens the ranking.
+    pub rrf_k: f64,
+    /// Weight of the recency signal relative to vector similarity. Damped on
+    /// purpose: at parity the newest rows occupy every slot no matter what was
+    /// asked. Raising it re-opens that.
+    pub recency_weight: f64,
+    /// MMR relevance/diversity trade-off, `0.0..=1.0`. Higher favors relevance.
+    pub mmr_lambda: f32,
+    /// Mean top-k cosine below which recall falls back to labeled lexical
+    /// search rather than dressing weak hits up as grounding.
+    pub min_coverage: f32,
+    /// How many top vector hits define that coverage estimate.
+    pub coverage_topk: usize,
+    /// Graph-expansion seeds taken from the strongest vector hits.
+    pub max_vector_seeds: usize,
+    /// Cap on lexical-fallback frames.
+    pub lexical_limit: usize,
+    /// Shortlist size as a multiple of `max_frames` — what the budget chooses
+    /// between, and the denominator of the drop report.
+    pub mmr_candidate_multiple: usize,
+}
+
+impl Default for RetrievalSettings {
+    fn default() -> Self {
+        let t = stella_context::RecallTuning::default();
+        Self {
+            // The per-query budgets, previously literals at the one call site
+            // that builds a prompt recall.
+            max_frames: DEFAULT_RECALL_MAX_FRAMES,
+            max_tokens: DEFAULT_RECALL_MAX_TOKENS,
+            rrf_k: t.rrf_k,
+            recency_weight: t.recency_weight,
+            mmr_lambda: t.mmr_lambda,
+            min_coverage: t.min_coverage,
+            coverage_topk: t.coverage_topk,
+            max_vector_seeds: t.max_vector_seeds,
+            lexical_limit: t.lexical_limit,
+            mmr_candidate_multiple: t.mmr_candidate_multiple,
+        }
+    }
+}
+
+/// Frames a prompt recall asks for — the value that shipped as a literal.
+pub const DEFAULT_RECALL_MAX_FRAMES: u32 = 5;
+/// Token budget those frames are packed into — likewise.
+pub const DEFAULT_RECALL_MAX_TOKENS: u32 = 1200;
+
+impl RetrievalSettings {
+    /// The store-level knobs, as the context plane's own type.
+    ///
+    /// `max_frames`/`max_tokens` are deliberately not here: they are per-query,
+    /// travel on `ContextQuery`, and a store-wide copy of them would be a
+    /// second place for the same number to live.
+    #[must_use]
+    pub fn tuning(&self) -> stella_context::RecallTuning {
+        stella_context::RecallTuning {
+            rrf_k: self.rrf_k,
+            recency_weight: self.recency_weight,
+            mmr_lambda: self.mmr_lambda,
+            min_coverage: self.min_coverage,
+            coverage_topk: self.coverage_topk,
+            max_vector_seeds: self.max_vector_seeds,
+            lexical_limit: self.lexical_limit,
+            mmr_candidate_multiple: self.mmr_candidate_multiple,
+        }
+        .sanitized()
+    }
+}
+
 #[cfg(test)]
 mod tests {
     use super::*;
@@ -296,6 +385,18 @@ mod tests {
                 "raw_observation_days":15,
                 "proposal_days":45,
                 "inferred_directive_review_days":200
+            },
+            "retrieval":{
+                "max_frames":11,
+                "max_tokens":2222,
+                "rrf_k":33.0,
+                "recency_weight":0.44,
+                "mmr_lambda":0.55,
+                "min_coverage":0.66,
+                "coverage_topk":7,
+                "max_vector_seeds":13,
+                "lexical_limit":17,
+                "mmr_candidate_multiple":6
             }
         }}"#;
         let s: Settings = serde_json::from_str(json).expect("parse");
@@ -327,5 +428,74 @@ mod tests {
         assert_eq!(ctx.retention.raw_observation_days, 15);
         assert_eq!(ctx.retention.proposal_days, 45);
         assert_eq!(ctx.retention.inferred_directive_review_days, 200);
+
+        let r = &ctx.retrieval;
+        assert_eq!(r.max_frames, 11);
+        assert_eq!(r.max_tokens, 2222);
+        assert_eq!(r.rrf_k, 33.0);
+        assert_eq!(r.recency_weight, 0.44);
+        assert_eq!(r.mmr_lambda, 0.55);
+        assert_eq!(r.min_coverage, 0.66);
+        assert_eq!(r.coverage_topk, 7);
+        assert_eq!(r.max_vector_seeds, 13);
+        assert_eq!(r.lexical_limit, 17);
+        assert_eq!(r.mmr_candidate_multiple, 6);
+    }
+
+    /// The defaults must reproduce the values that shipped as `const`s, or
+    /// "wire them, keeping current values as defaults" is not what happened.
+    #[test]
+    fn retrieval_defaults_are_the_constants_they_replaced() {
+        let d = RetrievalSettings::default();
+        assert_eq!(
+            d.max_frames, 5,
+            "the literal at the prompt-recall call site"
+        );
+        assert_eq!(d.max_tokens, 1200, "likewise");
+        assert_eq!(d.rrf_k, stella_context::DEFAULT_RRF_K);
+        assert_eq!(d.recency_weight, stella_context::DEFAULT_RECENCY_WEIGHT);
+        assert_eq!(d.mmr_lambda, stella_context::DEFAULT_MMR_LAMBDA);
+        assert_eq!(d.min_coverage, stella_context::DEFAULT_MIN_COVERAGE);
+        assert_eq!(d.coverage_topk, stella_context::DEFAULT_COVERAGE_TOPK);
+        assert_eq!(d.max_vector_seeds, stella_context::DEFAULT_MAX_VECTOR_SEEDS);
+        assert_eq!(d.lexical_limit, stella_context::DEFAULT_LEXICAL_LIMIT);
+        assert_eq!(
+            d.mmr_candidate_multiple,
+            stella_context::DEFAULT_MMR_CANDIDATE_MULTIPLE
+        );
+        assert_eq!(
+            d.tuning(),
+            stella_context::RecallTuning::default(),
+            "an unconfigured workspace hands the plane exactly its own defaults"
+        );
+    }
+
+    /// A nonsense value degrades retrieval; it must never break it. This is a
+    /// file a person edits, so zero and negative are reachable.
+    #[test]
+    fn out_of_range_tuning_is_clamped_not_rejected() {
+        let json = r#"{"context":{"retrieval":{
+            "rrf_k":-1.0,
+            "mmr_lambda":9.0,
+            "min_coverage":-4.0,
+            "coverage_topk":0,
+            "lexical_limit":0,
+            "mmr_candidate_multiple":0
+        }}}"#;
+        let s: Settings = serde_json::from_str(json).expect("parse");
+        let t = s.context.expect("present").retrieval.tuning();
+        assert_eq!(
+            t.rrf_k,
+            stella_context::DEFAULT_RRF_K,
+            "a non-positive fusion constant inverts the ranking"
+        );
+        assert_eq!(t.mmr_lambda, 1.0, "clamped into [0,1]");
+        assert_eq!(t.min_coverage, 0.0);
+        assert_eq!(t.coverage_topk, 1, "zero would divide by zero");
+        assert_eq!(t.lexical_limit, 1);
+        assert_eq!(
+            t.mmr_candidate_multiple, 1,
+            "zero would make every recall empty"
+        );
     }
 }
